requirements.txt
torch>=2.0
numpy>=1.24
opencv-python>=4.7
Pillow>=9.0
matplotlib>=3.7
